backend/requirements.txt
langchain==0.1.5
langchainhub==0.1.14
langchain_openai==0.0.5
langchain-core==0.1.18
langchain-community==0.0.17
langserve==0.0.43
argparse==1.4.0
python-dotenv==1.0.0
chromadb==0.4.22
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
libmagic==1.0
openai==1.13.3
pdfminer.six==20231228
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
python-magic==0.4.27
pytz==2024.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
pypdf==4.1.0
